+++ internal/client/client.gen.go
@@ -26,6 +26,27 @@ const (
 	Outstanding ActionV1Status = "outstanding"
 )
 
+// Defines values for CatalogResourceV2Category.
+const (
+	CatalogResourceV2CategoryCustom    CatalogResourceV2Category = "custom"
+	CatalogResourceV2CategoryExternal  CatalogResourceV2Category = "external"
+	CatalogResourceV2CategoryPrimitive CatalogResourceV2Category = "primitive"
+)
+
+// Defines values for CatalogTypeAttributePayloadV2Mode.
+const (
+	CatalogTypeAttributePayloadV2ModeEmpty    CatalogTypeAttributePayloadV2Mode = ""
+	CatalogTypeAttributePayloadV2ModeExternal CatalogTypeAttributePayloadV2Mode = "external"
+	CatalogTypeAttributePayloadV2ModeManual   CatalogTypeAttributePayloadV2Mode = "manual"
+)
+
+// Defines values for CatalogTypeAttributeV2Mode.
+const (
+	Empty    CatalogTypeAttributeV2Mode = ""
+	External CatalogTypeAttributeV2Mode = "external"
+	Manual   CatalogTypeAttributeV2Mode = "manual"
+)
+
 // Defines values for CatalogTypeV2Color.
 const (
 	CatalogTypeV2ColorBlue   CatalogTypeV2Color = "blue"
@@ -86,8 +107,9 @@ const (
 
 // Defines values for CreateRequestBody5Category.
 const (
-	CreateRequestBody5CategoryClosed CreateRequestBody5Category = "closed"
-	CreateRequestBody5CategoryLive   CreateRequestBody5Category = "live"
+	CreateRequestBody5CategoryClosed   CreateRequestBody5Category = "closed"
+	CreateRequestBody5CategoryLearning CreateRequestBody5Category = "learning"
+	CreateRequestBody5CategoryLive     CreateRequestBody5Category = "live"
 )
 
 // Defines values for CreateRequestBody6Mode.
@@ -215,6 +237,8 @@ const (
 
 // Defines values for IdentityV1Roles.
 const (
+	IdentityV1RolesCatalogEditor   IdentityV1Roles = "catalog_editor"
+	IdentityV1RolesCatalogViewer   IdentityV1Roles = "catalog_viewer"
 	IdentityV1RolesGlobalAccess    IdentityV1Roles = "global_access"
 	IdentityV1RolesIncidentCreator IdentityV1Roles = "incident_creator"
 	IdentityV1RolesIncidentEditor  IdentityV1Roles = "incident_editor"
@@ -231,8 +255,10 @@ const (
 
 // Defines values for IncidentStatusV1Category.
 const (
+	IncidentStatusV1CategoryCanceled IncidentStatusV1Category = "canceled"
 	IncidentStatusV1CategoryClosed   IncidentStatusV1Category = "closed"
 	IncidentStatusV1CategoryDeclined IncidentStatusV1Category = "declined"
+	IncidentStatusV1CategoryLearning IncidentStatusV1Category = "learning"
 	IncidentStatusV1CategoryLive     IncidentStatusV1Category = "live"
 	IncidentStatusV1CategoryMerged   IncidentStatusV1Category = "merged"
 	IncidentStatusV1CategoryTriage   IncidentStatusV1Category = "triage"
@@ -286,6 +312,36 @@ const (
 	UpdateRequestBody2RequiredAlways        UpdateRequestBody2Required = "always"
 	UpdateRequestBody2RequiredBeforeClosure UpdateRequestBody2Required = "before_closure"
 	UpdateRequestBody2RequiredNever         UpdateRequestBody2Required = "never"
+)
+
+// Defines values for UpdateTypeRequestBodyColor.
+const (
+	Blue   UpdateTypeRequestBodyColor = "blue"
+	Green  UpdateTypeRequestBodyColor = "green"
+	Red    UpdateTypeRequestBodyColor = "red"
+	Slate  UpdateTypeRequestBodyColor = "slate"
+	Violet UpdateTypeRequestBodyColor = "violet"
+	Yellow UpdateTypeRequestBodyColor = "yellow"
+)
+
+// Defines values for UpdateTypeRequestBodyIcon.
+const (
+	Bolt      UpdateTypeRequestBodyIcon = "bolt"
+	Box       UpdateTypeRequestBodyIcon = "box"
+	Briefcase UpdateTypeRequestBodyIcon = "briefcase"
+	Browser   UpdateTypeRequestBodyIcon = "browser"
+	Bulb      UpdateTypeRequestBodyIcon = "bulb"
+	Clock     UpdateTypeRequestBodyIcon = "clock"
+	Cog       UpdateTypeRequestBodyIcon = "cog"
+	Database  UpdateTypeRequestBodyIcon = "database"
+	Doc       UpdateTypeRequestBodyIcon = "doc"
+	Email     UpdateTypeRequestBodyIcon = "email"
+	Server    UpdateTypeRequestBodyIcon = "server"
+	Severity  UpdateTypeRequestBodyIcon = "severity"
+	Star      UpdateTypeRequestBodyIcon = "star"
+	Tag       UpdateTypeRequestBodyIcon = "tag"
+	User      UpdateTypeRequestBodyIcon = "user"
+	Users     UpdateTypeRequestBodyIcon = "users"
 )
 
 // Defines values for UserV1Role.
@@ -421,6 +477,9 @@ type CatalogEntryV2 struct {
 	// Aliases Optional aliases that can be used to reference this entry
 	Aliases []string `json:"aliases"`
 
+	// ArchivedAt When this entry was archived
+	ArchivedAt *time.Time `json:"archived_at,omitempty"`
+
 	// AttributeValues Values of this entry
 	AttributeValues map[string]CatalogAttributeBindingV2 `json:"attribute_values"`
 
@@ -446,6 +505,27 @@ type CatalogEntryV2 struct {
 	UpdatedAt time.Time `json:"updated_at"`
 }
 
+// CatalogResourceV2 defines model for CatalogResourceV2.
+type CatalogResourceV2 struct {
+	// Category Which category of resource
+	Category CatalogResourceV2Category `json:"category"`
+
+	// Description Human readable description for this resource
+	Description string `json:"description"`
+
+	// Label Label for this catalog resource type
+	Label string `json:"label"`
+
+	// Type Catalog type name for this resource
+	Type string `json:"type"`
+
+	// ValueDocstring Documentation for the literal string value of this resource
+	ValueDocstring string `json:"value_docstring"`
+}
+
+// CatalogResourceV2Category Which category of resource
+type CatalogResourceV2Category string
+
 // CatalogTypeAttributePayloadV2 defines model for CatalogTypeAttributePayloadV2.
 type CatalogTypeAttributePayloadV2 struct {
 	// Array Whether this attribute is an array
@@ -454,12 +534,18 @@ type CatalogTypeAttributePayloadV2 struct {
 	// Id The ID of this attribute
 	Id *string `json:"id,omitempty"`
 
+	// Mode Controls how this attribute is modified
+	Mode *CatalogTypeAttributePayloadV2Mode `json:"mode,omitempty"`
+
 	// Name Unique name of this attribute
 	Name string `json:"name"`
 
 	// Type Catalog type name for this attribute
 	Type string `json:"type"`
 }
+
+// CatalogTypeAttributePayloadV2Mode Controls how this attribute is modified
+type CatalogTypeAttributePayloadV2Mode string
 
 // CatalogTypeAttributeV2 defines model for CatalogTypeAttributeV2.
 type CatalogTypeAttributeV2 struct {
@@ -469,12 +555,18 @@ type CatalogTypeAttributeV2 struct {
 	// Id The ID of this attribute
 	Id string `json:"id"`
 
+	// Mode Controls how this attribute is modified
+	Mode CatalogTypeAttributeV2Mode `json:"mode"`
+
 	// Name Unique name of this attribute
 	Name string `json:"name"`
 
 	// Type Catalog type name for this attribute
 	Type string `json:"type"`
 }
+
+// CatalogTypeAttributeV2Mode Controls how this attribute is modified
+type CatalogTypeAttributeV2Mode string
 
 // CatalogTypeSchemaV2 defines model for CatalogTypeSchemaV2.
 type CatalogTypeSchemaV2 struct {
@@ -487,6 +579,9 @@ type CatalogTypeSchemaV2 struct {
 
 // CatalogTypeV2 defines model for CatalogTypeV2.
 type CatalogTypeV2 struct {
+	// Annotations Annotations that can track metadata about this type
+	Annotations map[string]string `json:"annotations"`
+
 	// Color Sets the display color of this type in the dashboard
 	Color CatalogTypeV2Color `json:"color"`
 
@@ -524,7 +619,7 @@ type CatalogTypeV2 struct {
 	// SemanticType Semantic type of this resource
 	SemanticType string `json:"semantic_type"`
 
-	// TypeName The type name of this catalog type, to be used when defining attributes
+	// TypeName The type name of this catalog type, to be used when defining attributes. This is immutable once a CatalogType has been created. For non-externally sync types, it must follow the pattern Custom["SomeName "]
 	TypeName string `json:"type_name"`
 
 	// UpdatedAt When this type was last updated
@@ -776,6 +871,9 @@ type CreateResponseBody struct {
 
 // CreateTypeRequestBody defines model for CreateTypeRequestBody.
 type CreateTypeRequestBody struct {
+	// Annotations Annotations that can track metadata about this type
+	Annotations *map[string]string `json:"annotations,omitempty"`
+
 	// Color Sets the display color of this type in the dashboard
 	Color *CreateTypeRequestBodyColor `json:"color,omitempty"`
 
@@ -793,6 +891,9 @@ type CreateTypeRequestBody struct {
 
 	// SemanticType Semantic type of this resource
 	SemanticType *string `json:"semantic_type,omitempty"`
+
+	// TypeName The type name of this catalog type, to be used when defining attributes. This is immutable once a CatalogType has been created. For non-externally sync types, it must follow the pattern Custom["SomeName "]
+	TypeName *string `json:"type_name,omitempty"`
 }
 
 // CreateTypeRequestBodyColor Sets the display color of this type in the dashboard
@@ -811,7 +912,7 @@ type CustomFieldEntryPayloadV1 struct {
 	// CustomFieldId ID of the custom field this entry is linked against
 	CustomFieldId string `json:"custom_field_id"`
 
-	// Values List of values to associate with this entry
+	// Values List of values to associate with this entry. Use an empty array to unset the value of the custom field.
 	Values []CustomFieldValuePayloadV1 `json:"values"`
 }
 
@@ -1333,6 +1434,18 @@ type IncidentV2 struct {
 
 	// Visibility Whether the incident should be open to anyone in your Slack workspace (public), or invite-only (private). For more information on Private Incidents see our [help centre](https://help.incident.io/en/articles/5947963-can-we-mark-incidents-as-sensitive-and-restrict-access).
 	Visibility IncidentV2Visibility `json:"visibility"`
+
+	// WorkloadMinutesLate Amount of time spent on the incident in late hours
+	WorkloadMinutesLate *float64 `json:"workload_minutes_late,omitempty"`
+
+	// WorkloadMinutesSleeping Amount of time spent on the incident in sleeping hours
+	WorkloadMinutesSleeping *float64 `json:"workload_minutes_sleeping,omitempty"`
+
+	// WorkloadMinutesTotal Amount of time spent on the incident in total
+	WorkloadMinutesTotal *float64 `json:"workload_minutes_total,omitempty"`
+
+	// WorkloadMinutesWorking Amount of time spent on the incident in working hours
+	WorkloadMinutesWorking *float64 `json:"workload_minutes_working,omitempty"`
 }
 
 // IncidentV2Mode Whether the incident is real, a test, a tutorial, or importing as a retrospective incident
@@ -1343,9 +1456,14 @@ type IncidentV2Visibility string
 
 // ListEntriesResponseBody defines model for ListEntriesResponseBody.
 type ListEntriesResponseBody struct {
-	CatalogEntries []CatalogEntryV2 `json:"catalog_entries"`
-	CatalogType    CatalogTypeV2    `json:"catalog_type"`
-	PaginationMeta PaginationMeta   `json:"pagination_meta"`
+	CatalogEntries []CatalogEntryV2     `json:"catalog_entries"`
+	CatalogType    CatalogTypeV2        `json:"catalog_type"`
+	PaginationMeta PaginationMetaResult `json:"pagination_meta"`
+}
+
+// ListResourcesResponseBody defines model for ListResourcesResponseBody.
+type ListResourcesResponseBody struct {
+	Resources []CatalogResourceV2 `json:"resources"`
 }
 
 // ListResponseBody defines model for ListResponseBody.
@@ -1355,14 +1473,14 @@ type ListResponseBody struct {
 
 // ListResponseBody10 defines model for ListResponseBody10.
 type ListResponseBody10 struct {
-	Incidents      []IncidentV1             `json:"incidents"`
-	PaginationMeta *PaginationMetaWithTotal `json:"pagination_meta,omitempty"`
+	Incidents      []IncidentV1                   `json:"incidents"`
+	PaginationMeta *PaginationMetaResultWithTotal `json:"pagination_meta,omitempty"`
 }
 
 // ListResponseBody11 defines model for ListResponseBody11.
 type ListResponseBody11 struct {
-	Incidents      []IncidentV2             `json:"incidents"`
-	PaginationMeta *PaginationMetaWithTotal `json:"pagination_meta,omitempty"`
+	Incidents      []IncidentV2                   `json:"incidents"`
+	PaginationMeta *PaginationMetaResultWithTotal `json:"pagination_meta,omitempty"`
 }
 
 // ListResponseBody12 defines model for ListResponseBody12.
@@ -1373,6 +1491,7 @@ type ListResponseBody12 struct {
 // ListResponseBody2 defines model for ListResponseBody2.
 type ListResponseBody2 struct {
 	CustomFieldOptions []CustomFieldOptionV1 `json:"custom_field_options"`
+	PaginationMeta     PaginationMetaResult  `json:"pagination_meta"`
 }
 
 // ListResponseBody3 defines model for ListResponseBody3.
@@ -1407,8 +1526,8 @@ type ListResponseBody8 struct {
 
 // ListResponseBody9 defines model for ListResponseBody9.
 type ListResponseBody9 struct {
-	IncidentUpdates []IncidentUpdateV2 `json:"incident_updates"`
-	PaginationMeta  *PaginationMeta    `json:"pagination_meta,omitempty"`
+	IncidentUpdates []IncidentUpdateV2    `json:"incident_updates"`
+	PaginationMeta  *PaginationMetaResult `json:"pagination_meta,omitempty"`
 }
 
 // ListTypesResponseBody defines model for ListTypesResponseBody.
@@ -1416,18 +1535,18 @@ type ListTypesResponseBody struct {
 	CatalogTypes []CatalogTypeV2 `json:"catalog_types"`
 }
 
-// PaginationMeta defines model for PaginationMeta.
-type PaginationMeta struct {
-	// After If provided, were records after a particular ID
+// PaginationMetaResult defines model for PaginationMetaResult.
+type PaginationMetaResult struct {
+	// After If provided, pass this as the 'after' param to load the next page
 	After *string `json:"after,omitempty"`
 
 	// PageSize What was the maximum number of results requested
 	PageSize int64 `json:"page_size"`
 }
 
-// PaginationMetaWithTotal defines model for PaginationMetaWithTotal.
-type PaginationMetaWithTotal struct {
-	// After If provided, were records after a particular ID
+// PaginationMetaResultWithTotal defines model for PaginationMetaResultWithTotal.
+type PaginationMetaResultWithTotal struct {
+	// After If provided, pass this as the 'after' param to load the next page
 	After *string `json:"after,omitempty"`
 
 	// PageSize What was the maximum number of results requested
@@ -1601,6 +1720,36 @@ type UpdateRequestBody4 struct {
 	Name string `json:"name"`
 }
 
+// UpdateTypeRequestBody defines model for UpdateTypeRequestBody.
+type UpdateTypeRequestBody struct {
+	// Annotations Annotations that can track metadata about this type
+	Annotations *map[string]string `json:"annotations,omitempty"`
+
+	// Color Sets the display color of this type in the dashboard
+	Color *UpdateTypeRequestBodyColor `json:"color,omitempty"`
+
+	// Description Human readble description of this type
+	Description string `json:"description"`
+
+	// Icon Sets the display icon of this type in the dashboard
+	Icon *UpdateTypeRequestBodyIcon `json:"icon,omitempty"`
+
+	// Name Name is the human readable name of this type
+	Name string `json:"name"`
+
+	// Ranked If this type should be ranked
+	Ranked *bool `json:"ranked,omitempty"`
+
+	// SemanticType Semantic type of this resource
+	SemanticType *string `json:"semantic_type,omitempty"`
+}
+
+// UpdateTypeRequestBodyColor Sets the display color of this type in the dashboard
+type UpdateTypeRequestBodyColor string
+
+// UpdateTypeRequestBodyIcon Sets the display icon of this type in the dashboard
+type UpdateTypeRequestBodyIcon string
+
 // UpdateTypeSchemaRequestBody defines model for UpdateTypeSchemaRequestBody.
 type UpdateTypeSchemaRequestBody struct {
 	Attributes []CatalogTypeAttributePayloadV2 `json:"attributes"`
@@ -1685,7 +1834,7 @@ type IncidentAttachmentsV1ListParamsResourceType string
 // IncidentsV1ListParams defines parameters for IncidentsV1List.
 type IncidentsV1ListParams struct {
 	// PageSize Integer number of records to return
-	PageSize *int64 `form:"page_size,omitempty" json:"page_size,omitempty"`
+	PageSize *int `form:"page_size,omitempty" json:"page_size,omitempty"`
 
 	// After An record's ID. This endpoint will return a list of records after this ID in relation to the API response order.
 	After *string `form:"after,omitempty" json:"after,omitempty"`
@@ -1700,7 +1849,7 @@ type CatalogV2ListEntriesParams struct {
 	CatalogTypeId string `form:"catalog_type_id" json:"catalog_type_id"`
 
 	// PageSize Integer number of records to return
-	PageSize *int64 `form:"page_size,omitempty" json:"page_size,omitempty"`
+	PageSize *int `form:"page_size,omitempty" json:"page_size,omitempty"`
 
 	// After An record's ID. This endpoint will return a list of records after this ID in relation to the API response order.
 	After *string `form:"after,omitempty" json:"after,omitempty"`
@@ -1712,7 +1861,7 @@ type IncidentUpdatesV2ListParams struct {
 	IncidentId *string `form:"incident_id,omitempty" json:"incident_id,omitempty"`
 
 	// PageSize Integer number of records to return
-	PageSize *int64 `form:"page_size,omitempty" json:"page_size,omitempty"`
+	PageSize *int `form:"page_size,omitempty" json:"page_size,omitempty"`
 
 	// After An record's ID. This endpoint will return a list of records after this ID in relation to the API response order.
 	After *string `form:"after,omitempty" json:"after,omitempty"`
@@ -1788,7 +1937,7 @@ type CatalogV2UpdateEntryJSONRequestBody = UpdateEntryRequestBody
 type CatalogV2CreateTypeJSONRequestBody = CreateTypeRequestBody
 
 // CatalogV2UpdateTypeJSONRequestBody defines body for CatalogV2UpdateType for application/json ContentType.
-type CatalogV2UpdateTypeJSONRequestBody = CreateTypeRequestBody
+type CatalogV2UpdateTypeJSONRequestBody = UpdateTypeRequestBody
 
 // CatalogV2UpdateTypeSchemaJSONRequestBody defines body for CatalogV2UpdateTypeSchema for application/json ContentType.
 type CatalogV2UpdateTypeSchemaJSONRequestBody = UpdateTypeSchemaRequestBody
@@ -2025,6 +2174,9 @@ type ClientInterface interface {
 	CatalogV2UpdateEntryWithBody(ctx context.Context, id string, contentType string, body io.Reader, reqEditors ...RequestEditorFn) (*http.Response, error)
 
 	CatalogV2UpdateEntry(ctx context.Context, id string, body CatalogV2UpdateEntryJSONRequestBody, reqEditors ...RequestEditorFn) (*http.Response, error)
+
+	// CatalogV2ListResources request
+	CatalogV2ListResources(ctx context.Context, reqEditors ...RequestEditorFn) (*http.Response, error)
 
 	// CatalogV2ListTypes request
 	CatalogV2ListTypes(ctx context.Context, reqEditors ...RequestEditorFn) (*http.Response, error)
@@ -2738,6 +2890,18 @@ func (c *Client) CatalogV2UpdateEntryWithBody(ctx context.Context, id string, co
 
 func (c *Client) CatalogV2UpdateEntry(ctx context.Context, id string, body CatalogV2UpdateEntryJSONRequestBody, reqEditors ...RequestEditorFn) (*http.Response, error) {
 	req, err := NewCatalogV2UpdateEntryRequest(c.Server, id, body)
+	if err != nil {
+		return nil, err
+	}
+	req = req.WithContext(ctx)
+	if err := c.applyEditors(ctx, req, reqEditors); err != nil {
+		return nil, err
+	}
+	return c.Client.Do(req)
+}
+
+func (c *Client) CatalogV2ListResources(ctx context.Context, reqEditors ...RequestEditorFn) (*http.Response, error) {
+	req, err := NewCatalogV2ListResourcesRequest(c.Server)
 	if err != nil {
 		return nil, err
 	}
@@ -4686,6 +4850,33 @@ func NewCatalogV2UpdateEntryRequestWithBody(server string, id string, contentTyp
 	return req, nil
 }
 
+// NewCatalogV2ListResourcesRequest generates requests for CatalogV2ListResources
+func NewCatalogV2ListResourcesRequest(server string) (*http.Request, error) {
+	var err error
+
+	serverURL, err := url.Parse(server)
+	if err != nil {
+		return nil, err
+	}
+
+	operationPath := fmt.Sprintf("/v2/catalog_resources")
+	if operationPath[0] == '/' {
+		operationPath = "." + operationPath
+	}
+
+	queryURL, err := serverURL.Parse(operationPath)
+	if err != nil {
+		return nil, err
+	}
+
+	req, err := http.NewRequest("GET", queryURL.String(), nil)
+	if err != nil {
+		return nil, err
+	}
+
+	return req, nil
+}
+
 // NewCatalogV2ListTypesRequest generates requests for CatalogV2ListTypes
 func NewCatalogV2ListTypesRequest(server string) (*http.Request, error) {
 	var err error
@@ -5515,6 +5706,9 @@ type ClientWithResponsesInterface interface {
 	CatalogV2UpdateEntryWithBodyWithResponse(ctx context.Context, id string, contentType string, body io.Reader, reqEditors ...RequestEditorFn) (*CatalogV2UpdateEntryResponse, error)
 
 	CatalogV2UpdateEntryWithResponse(ctx context.Context, id string, body CatalogV2UpdateEntryJSONRequestBody, reqEditors ...RequestEditorFn) (*CatalogV2UpdateEntryResponse, error)
+
+	// CatalogV2ListResources request
+	CatalogV2ListResourcesWithResponse(ctx context.Context, reqEditors ...RequestEditorFn) (*CatalogV2ListResourcesResponse, error)
 
 	// CatalogV2ListTypes request
 	CatalogV2ListTypesWithResponse(ctx context.Context, reqEditors ...RequestEditorFn) (*CatalogV2ListTypesResponse, error)
@@ -6483,6 +6677,28 @@ func (r CatalogV2UpdateEntryResponse) StatusCode() int {
 	return 0
 }
 
+type CatalogV2ListResourcesResponse struct {
+	Body         []byte
+	HTTPResponse *http.Response
+	JSON200      *ListResourcesResponseBody
+}
+
+// Status returns HTTPResponse.Status
+func (r CatalogV2ListResourcesResponse) Status() string {
+	if r.HTTPResponse != nil {
+		return r.HTTPResponse.Status
+	}
+	return http.StatusText(0)
+}
+
+// StatusCode returns HTTPResponse.StatusCode
+func (r CatalogV2ListResourcesResponse) StatusCode() int {
+	if r.HTTPResponse != nil {
+		return r.HTTPResponse.StatusCode
+	}
+	return 0
+}
+
 type CatalogV2ListTypesResponse struct {
 	Body         []byte
 	HTTPResponse *http.Response
@@ -7256,6 +7472,15 @@ func (c *ClientWithResponses) CatalogV2UpdateEntryWithResponse(ctx context.Conte
 		return nil, err
 	}
 	return ParseCatalogV2UpdateEntryResponse(rsp)
+}
+
+// CatalogV2ListResourcesWithResponse request returning *CatalogV2ListResourcesResponse
+func (c *ClientWithResponses) CatalogV2ListResourcesWithResponse(ctx context.Context, reqEditors ...RequestEditorFn) (*CatalogV2ListResourcesResponse, error) {
+	rsp, err := c.CatalogV2ListResources(ctx, reqEditors...)
+	if err != nil {
+		return nil, err
+	}
+	return ParseCatalogV2ListResourcesResponse(rsp)
 }
 
 // CatalogV2ListTypesWithResponse request returning *CatalogV2ListTypesResponse
@@ -8427,6 +8652,32 @@ func ParseCatalogV2UpdateEntryResponse(rsp *http.Response) (*CatalogV2UpdateEntr
 	switch {
 	case strings.Contains(rsp.Header.Get("Content-Type"), "json") && rsp.StatusCode == 200:
 		var dest ShowEntryResponseBody
+		if err := json.Unmarshal(bodyBytes, &dest); err != nil {
+			return nil, err
+		}
+		response.JSON200 = &dest
+
+	}
+
+	return response, nil
+}
+
+// ParseCatalogV2ListResourcesResponse parses an HTTP response from a CatalogV2ListResourcesWithResponse call
+func ParseCatalogV2ListResourcesResponse(rsp *http.Response) (*CatalogV2ListResourcesResponse, error) {
+	bodyBytes, err := io.ReadAll(rsp.Body)
+	defer func() { _ = rsp.Body.Close() }()
+	if err != nil {
+		return nil, err
+	}
+
+	response := &CatalogV2ListResourcesResponse{
+		Body:         bodyBytes,
+		HTTPResponse: rsp,
+	}
+
+	switch {
+	case strings.Contains(rsp.Header.Get("Content-Type"), "json") && rsp.StatusCode == 200:
+		var dest ListResourcesResponseBody
 		if err := json.Unmarshal(bodyBytes, &dest); err != nil {
 			return nil, err
 		}